GlobalKey offset retrieval bug
