+++ example/lib/main.dart
@@ -41,26 +41,27 @@ class NodeEditorExampleScreenState extends State<NodeEditorExampleScreen> {
       () => NodePrototype(
         name: 'Add',
         inputs: [
-          Port(id: 'a', name: 'A', data: 0, color: Colors.red),
-          Port(id: 'b', name: 'B', data: 0, color: Colors.green),
+          PortPrototype(
+            name: 'A',
+            data: 0,
+          ),
+          PortPrototype(
+            name: 'B',
+            data: 0,
+          ),
         ],
         outputs: [
-          Port(id: 'result', name: 'Result', data: 0, color: Colors.blue),
+          PortPrototype(
+            name: 'Result',
+            data: 0,
+          ),
         ],
-        fields: [
-          Field(id: 'a', name: 'A', data: 0, color: Colors.red),
-          Field(id: 'b', name: 'B', data: 0, color: Colors.green),
-        ],
-        onExecute: (inputIds) {
-          final a = int.parse(inputIds[0]);
-          final b = int.parse(inputIds[1]);
-          return [(a + b).toString()];
-        },
-        color: Colors.blue,
+        onExecute: (inputIds, outputIds) {},
       ),
     );
 
-    _nodeEditorController.addNode('add');
+    _nodeEditorController.addNode('add', offset: const Offset(100, 100));
+    _nodeEditorController.addNode('add', offset: const Offset(300, 500));
   }
 
   @override
@@ -78,7 +79,7 @@ class NodeEditorExampleScreenState extends State<NodeEditorExampleScreen> {
         backgroundColor: Colors.blue,
       ),
       body: Center(
-        child: FlNodeEditorWidget(
+        child: FlNodeEditor(
           controller: _nodeEditorController,
           expandToParent: true,
           style: const NodeEditorStyle(
@@ -89,58 +90,63 @@ class NodeEditorExampleScreenState extends State<NodeEditorExampleScreen> {
               intersectionRadius: 2,
             ),
           ),
-          content: (offset, zoom) {
+          overaly: (offset, zoom) {
             return [
-              Column(
-                mainAxisAlignment: MainAxisAlignment.end,
-                crossAxisAlignment: CrossAxisAlignment.center,
-                children: [
-                  Container(
-                    decoration: BoxDecoration(
-                      color: Colors.blue,
-                      borderRadius: BorderRadius.circular(8),
-                    ),
-                    child: IconButton(
-                      onPressed: () =>
-                          _nodeEditorController.setOffset(Offset.zero),
-                      icon: const Icon(
-                        Icons.center_focus_strong,
-                        size: 32,
-                        color: Colors.white,
+              Padding(
+                padding: const EdgeInsets.all(8),
+                child: Column(
+                  mainAxisAlignment: MainAxisAlignment.end,
+                  crossAxisAlignment: CrossAxisAlignment.center,
+                  children: [
+                    Container(
+                      decoration: BoxDecoration(
+                        color: Colors.blue,
+                        borderRadius: BorderRadius.circular(8),
+                      ),
+                      child: IconButton(
+                        onPressed: () =>
+                            _nodeEditorController.setOffset(Offset.zero),
+                        icon: const Icon(
+                          Icons.center_focus_strong,
+                          size: 32,
+                          color: Colors.white,
+                        ),
                       ),
                     ),
-                  ),
-                  const Gap(8),
-                  Container(
-                    decoration: BoxDecoration(
-                      color: Colors.blue,
-                      borderRadius: BorderRadius.circular(8),
-                    ),
-                    child: IconButton(
-                      onPressed: () => _nodeEditorController.setZoom(zoom * 2),
-                      icon: const Icon(
-                        Icons.zoom_in,
-                        size: 32,
-                        color: Colors.white,
+                    const Gap(8),
+                    Container(
+                      decoration: BoxDecoration(
+                        color: Colors.blue,
+                        borderRadius: BorderRadius.circular(8),
+                      ),
+                      child: IconButton(
+                        onPressed: () =>
+                            _nodeEditorController.setZoom(zoom * 2),
+                        icon: const Icon(
+                          Icons.zoom_in,
+                          size: 32,
+                          color: Colors.white,
+                        ),
                       ),
                     ),
-                  ),
-                  const Gap(8),
-                  Container(
-                    decoration: BoxDecoration(
-                      color: Colors.blue,
-                      borderRadius: BorderRadius.circular(8),
-                    ),
-                    child: IconButton(
-                      onPressed: () => _nodeEditorController.setZoom(zoom / 2),
-                      icon: const Icon(
-                        Icons.zoom_out,
-                        size: 32,
-                        color: Colors.white,
+                    const Gap(8),
+                    Container(
+                      decoration: BoxDecoration(
+                        color: Colors.blue,
+                        borderRadius: BorderRadius.circular(8),
+                      ),
+                      child: IconButton(
+                        onPressed: () =>
+                            _nodeEditorController.setZoom(zoom / 2),
+                        icon: const Icon(
+                          Icons.zoom_out,
+                          size: 32,
+                          color: Colors.white,
+                        ),
                       ),
                     ),
-                  ),
-                ],
+                  ],
+                ),
               ),
             ];
           },
