+++ lib/fl_nodes.dart
@@ -3,4 +3,4 @@ library fl_nodes;
 export 'package:fl_nodes/src/core/models/styles.dart';
 export 'package:fl_nodes/src/core/models/node.dart';
 export 'package:fl_nodes/src/core/controllers/node_editor.dart';
-export 'package:fl_nodes/src/widgets/node_editor.dart';
+export 'package:fl_nodes/src/widgets/fl_nodes.dart';
