+++ lib/src/core/utils/math.dart
@@ -1,0 +1,5 @@
+import 'dart:math';
+
+num maxIn(List<num> numbers) {
+  return numbers.reduce((a, b) => max(a, b));
+}
